[PLAT-530]: show last refreshed as a duration
